pyproject.toml
[tool.poetry]
name = "ipif-client"
version = "0.1.0"
description = ""
authors = ["Richard Hadden <richard.hadden@oeaw.ac.at>"]

[tool.poetry.dependencies]
python = "^3.9"
requests = "^2.25.1"
pytest-httpserver = "^1.0.0"
bunch = "^1.0.1"
python-dateutil = "^2.8.1"
yaspin = "^2.0.0"
munch = "^2.5.0"
multi_key_dict = "^2.0.3"
ordered-set = "^4.0.2"
multilookupdict = "^0.1.2"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = {version = "^21.5b2", allow-prereleases = true}
ipython = "^7.24.1"
pytest-sugar = "^0.9.4"
pytest-mock = "^3.6.1"
jupyter = "^1.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
